Normalize test/suite name: hosted app copy verify
